Move: move defaultFrame to component package
